Reconfigured the data layer to dynamicaly build the db as you use the site.

+++ ThrowAwayProjects/Models/CarViewModel.cs
@@ -12,7 +12,7 @@ namespace ThrowAwayProjects.Models
         }
         public CarViewModel(Car _Car)
         {
-            Id = _Car.Id;
+            Id = (int)_Car.Id;
             Make = _Car.Make;
             Type = _Car.Model;
         }
